requirements.txt
flask
tensorflow
keras
scikit-learn
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
seaborn
matplotlib
pandas
numpy
